+++ messages/de/motion.php
@@ -136,6 +136,8 @@ Bitte beachte, dass sich die Unterstützer*innen deines Antrags mit ihren Zugang
     'withdraw_yes'                       => 'Zurückziehen',
     'withdrawn_adminnoti_title'          => 'Antrag zurückgezogen',
     'withdrawn_adminnoti_body'           => "Es wurde ein Antrag zurückgezogen.\nTitel: %TITLE%\nAntragsteller*in: %INITIATOR%\nLink: %LINK%",
+    'field_optional'                     => 'optional',
+    'field_required'                     => 'benötigt',
     'max_size_hint'                      => 'Max. %MB% MB',
     'max_len_hint'                       => 'Max. %LEN% Zeichen (Aktuell: %COUNT%)',
     'max_len_alert'                      => 'Der Text ist zu lang!',